+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Adding the type of component into the display of the stat module
 - Adding a flag to filter the type of components
+- Adding the number of drawings using the parts
 ### Changed
 ### Deprecated
 ### Removed